build_clang: Add support for BOLT optimization
from [1]: "BOLT is a post-link optimizer developed to speed up large applications. It achieves the improvements by optimizing application's code layout based on execution profile gathered by sampling profiler, such as Linux perf tool."

[1]: https://github.com/llvm/llvm-project/blob/main/bolt/README.md

+++ build_clang.sh
@@ -10,6 +10,7 @@ BINUTILS_VER="2_38"
 BUILDDIR=$(pwd)
 CLEAN_BUILD=3
 POLLY_OPT=1
+BOLT_OPT=1
 
 # DO NOT CHANGE
 USE_SYSTEM_BINUTILS_64=1
@@ -130,6 +131,160 @@ build_temp_binutils() {
 	make install -j$(nproc --all)
 }
 
+bolt_profile_gen() {
+
+	if [ "$1" = "perf" ]; then
+		echo "Training arm64"
+		cd "$KERNEL_DIR"
+		make distclean defconfig \
+			LLVM=1 \
+			LLVM_IAS=1 \
+			ARCH=arm64 \
+			CC="$STAGE3"/clang \
+			LD="$STAGE3"/ld.lld \
+			AR="$STAGE3"/llvm-ar \
+			NM="$STAGE3"/llvm-nm \
+			LD="$STAGE3"/ld.lld \
+			STRIP="$STAGE3"/llvm-strip \
+			OBJCOPY="$STAGE3"/llvm-objcopy \
+			OBJDUMP="$STAGE3"/llvm-objdump \
+			OBJSIZE="$STAGE3"/llvm-size \
+			HOSTCC="$STAGE3"/clang \
+			HOSTCXX="$STAGE3"/clang++ \
+			HOSTAR="$STAGE3"/llvm-ar \
+			HOSTLD="$STAGE3"/ld.lld \
+			CROSS_COMPILE=aarch64-linux-gnu-
+
+		perf record --output ${BOLT_PROFILES}/perf.data --event cycles:u --branch-filter any,u -- make all -s -j$(nproc --all) \
+			LLVM=1 \
+			LLVM_IAS=1 \
+			ARCH=arm64 \
+			CC="$STAGE3"/clang \
+			LD="$STAGE3"/ld.lld \
+			AR="$STAGE3"/llvm-ar \
+			NM="$STAGE3"/llvm-nm \
+			LD="$STAGE3"/ld.lld \
+			STRIP="$STAGE3"/llvm-strip \
+			OBJCOPY="$STAGE3"/llvm-objcopy \
+			OBJDUMP="$STAGE3"/llvm-objdump \
+			OBJSIZE="$STAGE3"/llvm-size \
+			HOSTCC="$STAGE3"/clang \
+			HOSTCXX="$STAGE3"/clang++ \
+			HOSTAR="$STAGE3"/llvm-ar \
+			HOSTLD="$STAGE3"/ld.lld \
+			CROSS_COMPILE=aarch64-linux-gnu- || (
+			echo "Kernel Build failed!"
+			exit 1
+		)
+		cd "$OUT"
+
+		"$STAGE1"/perf2bolt "$STAGE3"/clang-16 \
+			-p ${BOLT_PROFILES}/perf.data \
+			-o ${BOLT_PROFILES}/clang-16.fdata || (
+			echo "Failed to convert perf data"
+			exit 1
+		)
+
+		"$STAGE1"/llvm-bolt "$STAGE3"/clang-16 \
+			-o "$STAGE3"/clang-16.bolt \
+			--data ${BOLT_PROFILES}/clang-16.fdata \
+			-relocs \
+			-split-functions \
+			-split-all-cold \
+			-icf=1 \
+			-lite=1 \
+			-split-eh \
+			-use-gnu-stack \
+			-jump-tables=move \
+			-dyno-stats \
+			-reorder-functions=hfsort \
+			-reorder-blocks=ext-tsp \
+			-tail-duplication=cache || (
+			echo "Could not optimize clang with BOLT"
+			exit 1
+		)
+
+		mv "$STAGE3"/clang-16 "$STAGE3"/clang-16.org
+		mv "$STAGE3"/clang-16.bolt "$STAGE3"/clang-16
+	else
+		"$STAGE1"/llvm-bolt \
+			--instrument \
+			--instrumentation-file-append-pid \
+			--instrumentation-file=${BOLT_PROFILES}/clang-16.fdata \
+			"$STAGE3"/clang-16 \
+			-o "$STAGE3"/clang-16.inst
+
+		mv "$STAGE3"/clang-16 "$STAGE3"/clang-16.org
+		mv "$STAGE3"/clang-16.inst "$STAGE3"/clang-16
+
+		echo "Training arm64"
+		cd "$KERNEL_DIR"
+		make distclean defconfig \
+			LLVM=1 \
+			LLVM_IAS=1 \
+			ARCH=arm64 \
+			CC="$STAGE3"/clang \
+			LD="$STAGE3"/ld.lld \
+			AR="$STAGE3"/llvm-ar \
+			NM="$STAGE3"/llvm-nm \
+			LD="$STAGE3"/ld.lld \
+			STRIP="$STAGE3"/llvm-strip \
+			OBJCOPY="$STAGE3"/llvm-objcopy \
+			OBJDUMP="$STAGE3"/llvm-objdump \
+			OBJSIZE="$STAGE3"/llvm-size \
+			HOSTCC="$STAGE3"/clang \
+			HOSTCXX="$STAGE3"/clang++ \
+			HOSTAR="$STAGE3"/llvm-ar \
+			HOSTLD="$STAGE3"/ld.lld \
+			CROSS_COMPILE=aarch64-linux-gnu-
+
+		make all -s -j$(nproc --all) \
+			LLVM=1 \
+			LLVM_IAS=1 \
+			ARCH=arm64 \
+			CC="$STAGE3"/clang \
+			LD="$STAGE3"/ld.lld \
+			AR="$STAGE3"/llvm-ar \
+			NM="$STAGE3"/llvm-nm \
+			LD="$STAGE3"/ld.lld \
+			STRIP="$STAGE3"/llvm-strip \
+			OBJCOPY="$STAGE3"/llvm-objcopy \
+			OBJDUMP="$STAGE3"/llvm-objdump \
+			OBJSIZE="$STAGE3"/llvm-size \
+			HOSTCC="$STAGE3"/clang \
+			HOSTCXX="$STAGE3"/clang++ \
+			HOSTAR="$STAGE3"/llvm-ar \
+			HOSTLD="$STAGE3"/ld.lld \
+			CROSS_COMPILE=aarch64-linux-gnu- || (
+			echo "Kernel Build failed!"
+			exit 1
+		)
+		cd "$OUT"
+
+		cd $BOLT_PROFILES
+		"$STAGE1"/merge-fdata *.fdata >combined.fdata
+
+		"$STAGE1"/llvm-bolt "$STAGE3"/clang-16.org \
+			--data combined.fdata \
+			-o "$STAGE3"/clang-16 \
+			-relocs \
+			-split-functions \
+			-split-all-cold \
+			-icf=1 \
+			-lite=1 \
+			-split-eh \
+			-use-gnu-stack \
+			-jump-tables=move \
+			-dyno-stats \
+			-reorder-functions=hfsort \
+			-reorder-blocks=ext-tsp \
+			-tail-duplication=cache || (
+			echo "Could not optimize clang with BOLT"
+			exit 1
+		)
+	fi
+}
+
 if [ -d "$LLVM_DIR"/ ]; then
 	cd $LLVM_DIR/
 	if ! git status; then
@@ -189,7 +344,11 @@ OPT_FLAGS="-O3 -march=x86-64 -mtune=generic -ffunction-sections -fdata-sections"
 OPT_FLAGS_LD="-Wl,-O3,--sort-common,--as-needed,-z,now -fuse-ld=$LLVM_BIN_DIR/ld.lld"
 
 if [[ $POLLY_OPT -eq 1 ]]; then
-	STAGE1_PROJS="clang;lld;compiler-rt;polly"
+	if [[ $BOLT_OPT -eq 1 ]]; then
+		STAGE1_PROJS="clang;lld;compiler-rt;bolt;polly"
+	else
+		STAGE1_PROJS="clang;lld;compiler-rt;polly"
+	fi
 else
 	STAGE1_PROJS="clang;lld;compiler-rt"
 fi
@@ -434,8 +593,10 @@ time make all -s -j$(nproc --all) \
 cd "$PROFILES"
 "$STAGE2"/llvm-profdata merge -output=clang.profdata *
 
-rm -rf "$TEMP_BINTUILS_BUILD"
-rm -rf "$TEMP_BINTUILS_INSTALL"
+if [[ $BOLT_OPT -eq 0 ]]; then
+	rm -rf "$TEMP_BINTUILS_BUILD"
+	rm -rf "$TEMP_BINTUILS_INSTALL"
+fi
 
 echo "Stage 2: PGO Training End"
 
@@ -448,6 +609,12 @@ OPT_FLAGS="-march=x86-64 -mtune=generic -ffunction-sections -fdata-sections -flt
 
 if [[ $POLLY_OPT -eq 1 ]]; then
 	OPT_FLAGS="$OPT_FLAGS $POLLY_OPT_FLAGS"
+fi
+
+if [[ $BOLT_OPT -eq 1 ]]; then
+	OPT_FLAGS_LD_EXE="$OPT_FLAGS_LD -Wl,--emit-relocs"
+else
+	OPT_FLAGS_LD_EXE="$OPT_FLAGS_LD"
 fi
 
 cd "$LLVM_BUILD"
@@ -505,7 +672,7 @@ cmake -G Ninja -Wno-dev --log-level=NOTICE \
 	-DCMAKE_C_FLAGS="$OPT_FLAGS" \
 	-DCMAKE_ASM_FLAGS="$OPT_FLAGS" \
 	-DCMAKE_CXX_FLAGS="$OPT_FLAGS" \
-	-DCMAKE_EXE_LINKER_FLAGS="$OPT_FLAGS_LD" \
+	-DCMAKE_EXE_LINKER_FLAGS="$OPT_FLAGS_LD_EXE" \
 	-DCMAKE_MODULE_LINKER_FLAGS="$OPT_FLAGS_LD" \
 	-DCMAKE_SHARED_LINKER_FLAGS="$OPT_FLAGS_LD" \
 	-DCMAKE_INSTALL_PREFIX="$OUT/install" \
@@ -517,8 +684,32 @@ ninja install -j$(nproc --all) || (
 	exit 1
 )
 
-STAGE2="$OUT/install/bin"
+STAGE3="$OUT/install/bin"
 echo "Stage 3 Build: End"
+
+if [[ $BOLT_OPT -eq 1 ]]; then
+	# Optimize final built clang with BOLT
+	BOLT_PROFILES="$OUT/bolt-prof"
+	rm -rf $BOLT_PROFILES
+	mkdir -p "$BOLT_PROFILES"
+	export PATH="$STAGE3:$BINTUILS_64_BIN_DIR:$BINTUILS_32_BIN_DIR:$STOCK_PATH"
+	perf record -e cycles:u -j any,u -- sleep 1 &>/dev/null
+	if [[ $? == "0" ]]; then
+		echo "Performing BOLT with sampling!"
+		bolt_profile_gen "perf" || (
+			echo "Optimizing with BOLT failed!"
+			exit 1
+		)
+	else
+		echo "Performing BOLT with instrumenting!"
+		bolt_profile_gen "instrumenting" || (
+			echo "Optimizing with BOLT failed!"
+			exit 1
+		)
+	fi
+	rm -rf "$TEMP_BINTUILS_BUILD"
+	rm -rf "$TEMP_BINTUILS_INSTALL"
+fi
 
 echo "Moving stage 3 install dir to build dir"
 mv $OUT/install $BUILDDIR/install/
